added a cast, so wormux compiles again on x86_64 thanks ivze

+++ trunk/src/network/distant_cpu.cpp
@@ -230,7 +230,7 @@ void DistantComputer::ManageTeam(Action* team)
   {
     std::list<std::string>::iterator it;
     it = find(owned_teams.begin(), owned_teams.end(), name);
-    printf("size of owned teams: %d\n", owned_teams.size());
+    printf("size of owned teams: %d\n", (int)owned_teams.size());
 
     NET_ASSERT(it != owned_teams.end())
     {
